# swagger.yaml
openapi: 3.0.0
info:
  title: Battery Swapping And Health Monitoring API
  version: 1.0.0
paths:
  /api/biker/nearest-franchisers/{phoneNumber}:
    get:
      summary: Get nearest franchisers based on biker's phone number
      tags: [Biker]
      parameters:
        - in: path
          name: phoneNumber
          required: true
          description: Biker's phone number
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  nearestFranchisers:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: Name of the franchiser
                        coordinates:
                          type: array
                          items:
                            type: number
                          description: Latitude and longitude coordinates of the franchiser
                        distance:
                          type: number
                          description: Distance from the biker to the franchiser in kilometers
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
        '404':
          description: Biker not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
  /api/biker/requestSwap:
    post:
      summary: Create a new swap request
      tags: [Biker]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bikerPhoneNumber:
                  type: string
                franchiserPhoneNumber:
                  type: string
                batteryId:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Confirmation message
                  swapRequestId:
                    type: string
                    description: ID of the newly created swap request
        '404':
          description: Biker or franchiser not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
  /api/biker/checkSwapRequestStatus/{swapRequestId}:
    get:
      summary: Check swap request status
      tags: [Biker]
      parameters:
        - in: path
          name: swapRequestId
          required: true
          schema:
            type: string
          description: ID of the swap request
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Status message
        '404':
          description: Swap request not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Description of the error
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Description of the error
  /api/biker/register:
    post:
      summary: Register a biker
      tags: [Biker]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the biker
                email:
                  type: string
                  format: email
                  description: Email of the biker
                phoneNumber:
                  type: string
                  description: Phone number of the biker
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '404':
          description: Biker not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
  /api/biker/verify-email/{token}:
    get:
      summary: Verify biker's email
      tags: [Biker]
      parameters:
        - in: path
          name: token
          required: true
          description: JWT token sent to the biker's email
          schema:
            type: string
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '404':
          description: Biker not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
  /api/biker/check-email-verification:
    get:
      summary: Check email verification status
      tags: [Biker]
      parameters:
        - in: query
          name: email
          required: true
          description: Email of the biker to check
          schema:
            type: string
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '400':
          description: Email is required or not verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
        '404':
          description: Biker not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description of the error
...